pyproject.toml
[tool.poetry]
name = "poll-bot"
version = "0.1.0"
description = ""
authors = ["Fathur Rohman <hi.fathur.rohman@gmail.com>"]
readme = "README.md"
packages = [{include = "poll_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
python-dotenv = "^1.0.0"
pendulum = "^2.1.2"
nltk = "^3.8.1"
numpy = "^1.24.3"
masonite-orm = "^2.19.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
